ดึงหน้าชำระเงินได้แล้ว

+++ app/src/main/java/com/example/pchrp/newdashboard/Dao/objectdao/ObjectListItemDao.java
@@ -2,40 +2,40 @@ package com.example.pchrp.newdashboard.Dao.objectdao;
 
 public class ObjectListItemDao {
 
-    private Double cashPayments;
-    private Double creditPayments;
-    private Double creditCardPayments;
-    private Double totalIncome;
+    private Float cashPayments;
+    private Float creditPayments;
+    private Float creditCardPayments;
+    private Float totalIncome;
 
-    public Double getCashPayments() {
+    public Float getCashPayments() {
         return cashPayments;
     }
 
-    public void setCashPayments(Double cashPayments) {
+    public void setCashPayments(Float cashPayments) {
         this.cashPayments = cashPayments;
     }
 
-    public Double getCreditPayments() {
+    public Float getCreditPayments() {
         return creditPayments;
     }
 
-    public void setCreditPayments(Double creditPayments) {
+    public void setCreditPayments(Float creditPayments) {
         this.creditPayments = creditPayments;
     }
 
-    public Double getCreditCardPayments() {
+    public Float getCreditCardPayments() {
         return creditCardPayments;
     }
 
-    public void setCreditCardPayments(Double creditCardPayments) {
+    public void setCreditCardPayments(Float creditCardPayments) {
         this.creditCardPayments = creditCardPayments;
     }
 
-    public Double getTotalIncome() {
+    public Float getTotalIncome() {
         return totalIncome;
     }
 
-    public void setTotalIncome(Double totalIncome) {
+    public void setTotalIncome(Float totalIncome) {
         this.totalIncome = totalIncome;
     }
 }
